Add time, duraiton, and description to api call

+++ src/toggl_api.rs
@@ -50,7 +50,7 @@ pub struct TimeEntryRequest {
     pub created_with: String,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub description: Option<String>,
-    pub duration: i64,
+    pub duration: i32,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub duronly: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
